feat(6715): extract signer test modules
