+++ crates/kamn-node/src/main_tests.rs
@@ -257,5 +257,5 @@ mod report_tests;
 mod runtime_tests;
 mod service_api_endpoint_tests;
 mod service_api_endpoint_tests_split_contract;
-mod signer_tests_split_contract;
 mod signer_tests;
+mod signer_tests_split_contract;
