+++ crates/kamn-node/src/main_tests/signer_tests/signer_direct_profile_contract_tests.rs
@@ -1,0 +1,10 @@
+#[path = "signer_direct_profile_contract_tests/direct_signature_contract_tests.rs"]
+mod direct_signature_contract_tests;
+#[path = "signer_direct_profile_contract_tests/profile_parity_matrix_contract_tests.rs"]
+mod profile_parity_matrix_contract_tests;
+#[path = "signer_direct_profile_contract_tests/profile_selection_contract_tests.rs"]
+mod profile_selection_contract_tests;
+#[path = "signer_direct_profile_contract_tests/runtime_selection_contract_tests.rs"]
+mod runtime_selection_contract_tests;
+#[path = "signer_direct_profile_contract_tests/signer_source_policy_contract_tests.rs"]
+mod signer_source_policy_contract_tests;
