+++ crates/kamn-node/src/main_tests/signer_tests/signer_managed_external_contract_tests.rs
@@ -1,0 +1,18 @@
+#[path = "signer_managed_external_contract_tests/backend_command_contract_tests.rs"]
+mod backend_command_contract_tests;
+#[path = "signer_managed_external_contract_tests/backend_provenance_contract_tests.rs"]
+mod backend_provenance_contract_tests;
+#[path = "signer_managed_external_contract_tests/backend_provenance_parity_contract_tests.rs"]
+mod backend_provenance_parity_contract_tests;
+#[path = "signer_managed_external_contract_tests/backend_reason_code_contract_tests.rs"]
+mod backend_reason_code_contract_tests;
+#[path = "signer_managed_external_contract_tests/backend_required_marker_contract_tests.rs"]
+mod backend_required_marker_contract_tests;
+#[path = "signer_managed_external_contract_tests/backend_response_pubkey_contract_tests.rs"]
+mod backend_response_pubkey_contract_tests;
+#[path = "signer_managed_external_contract_tests/key_source_contract_tests.rs"]
+mod key_source_contract_tests;
+#[path = "signer_managed_external_contract_tests/runtime_pubkey_contract_tests.rs"]
+mod runtime_pubkey_contract_tests;
+#[path = "signer_managed_external_contract_tests/selection_contract_tests.rs"]
+mod selection_contract_tests;
