+++ crates/kamn-node/src/main_tests/signer_tests/signer_preflight_nonce_contract_tests.rs
@@ -1,0 +1,6 @@
+#[path = "signer_preflight_nonce_contract_tests/nonce_resolver_contract_tests.rs"]
+mod nonce_resolver_contract_tests;
+#[path = "signer_preflight_nonce_contract_tests/preflight_quorum_contract_tests.rs"]
+mod preflight_quorum_contract_tests;
+#[path = "signer_preflight_nonce_contract_tests/signer_preflight_policy_contract_tests.rs"]
+mod signer_preflight_policy_contract_tests;
